+++ packages/adapters/src/adapters/Cornerstone3D/RTStruct/utilities/getPatientModule.js
