+++ packages/adapters/src/adapters/Cornerstone3D/RTStruct/utilities/getRTSeriesModule.js
